4_evaluation/requirements.txt
pydantic
tqdm
python-dotenv
ruff
ragas
langchain-openai
langchain-core
openai
datasets
numpy
